Create functions to add and subtract input values
Removed function displayNumberString(). Defined variables numberString,
sumation, and deduction; defined arrays addArray and subtractArray.
When either the plus or minus button are pressed, the 'mousedown' event
moves the input value into the corresponding array (addArray or
subtractArray) and convert the input into a number. When the equal
button is pressed, the functions will either add or subtract all
elements within the corresponding array.