Fix operate function and condense condense
Replaced the enterNumber function with just the event listener for the
number buttons and the displayNumbers function to display "0" when no
number is entered or the numberString when it is not null. Removed
numberArray, calculateArray, and the pressedOperator function. Rewrote
operate function with an event listener for all operator buttons to
define the operator variable, and an event listener for the equal button
with a switch loop to perform the appropriate operation.