Fix the equal button and numberString issue
After a calculation was performed and the equal button was pressed, if a
number button was pressed, that button's text content was included to
the end of the numberString, which already was the result of the last
calculation. This was corrected by defining "result = null" before the
functions and event listeners and editing the code for the number button
event listener. Used the condition in which result was not null, both
the numberString and result variables would be result to their original
setting.