Fix value setting issues
Moved some code from different event listenters into the newly created
setValues and getResult functions to reuse the same code in other
situations as necessary. Added new conditions to the setValues function
to handle situations when the user tries to divide by 0 and then tries
to perform mathemathical functions on "JUST NO", which is not a number.